backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
requests
websockets